backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.32.3
gpxpy==1.6.2
pandas==2.2.3
numpy==2.1.2
cryptography==42.0.0
PyJWT==2.8.0
